ServiceWorker: Rewrite windowclient-navigate.html test.
This CL splits the test into each test case to find a failure reason easily.
Also, adds a test case for in-scope-but-not-controlled to the test.
The test case comes from the spec dicussion[1] and the comment[2] in other CL.

[1] https://github.com/slightlyoff/ServiceWorker/issues/752
[2] https://codereview.chromium.org/1481753002/#msg8

BUG=567049

Review URL: https://codereview.chromium.org/1604893002

Cr-Original-Commit-Position: refs/heads/master@{#375629}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 47a728652bd829f25edff357b5a640f9c78c328f