Make render pipeline throttling opt-in
Previously render pipeline throttling was allowed by default, and
subsystems which did not want throttling to happen needed to use
DocumentLifecycle::PreventThrottlingScope to turn it off. This was
error-prone and dangerous since it meant callers might see out-of-date
layout and style information unless they knew to opt out.

This patch flips the default to be that throttling is disallowed unless
DocumentLifecycle::AllowThrottlingScope is used to turn it on. We now
explicitly enable throttling for animations, lifecycle updates and hit
testing.

BUG=487937

Review URL: https://codereview.chromium.org/1696233002

Cr-Original-Commit-Position: refs/heads/master@{#376024}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 324897d72015abb00a5f7c682fbdc2e5d82619b0