Add origins argument to registerForeignFetchScopes.
Adds the new argument, stores the list of origins for which to intercept
requests, and adds origin checks to the interception code. Also updated the
various unit and layout tests.

This implements (parts of) the changes to the spec made in
https://github.com/mkruisselbrink/ServiceWorker/commit/88bbe9324fa8ac38d8af3e347564450d598757aa

BUG=540509

Review URL: https://codereview.chromium.org/1656933003

Cr-Original-Commit-Position: refs/heads/master@{#376036}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 0126c13fddde0066adf3ae782d135968b5ad8b6f