Keep track of accessibility focus across windows.
Previously we never kept track of accessibility focus; when a
web element or view got focus we'd fire an accessibility focus
event and leave it up to the client to keep track.

This worked fine when focus is only changing within a single
window. However, this led to lots of subtle bugs and race
conditions across windows, for example:

1. If a View in a non-activatable window tried to get focus,
we'd fire a focus event even though the View didn't actually
get keyboard focus.

2. If the active element in a non-focused frame changed,
we'd fire a focus event even though that frame's window
did not have focus.

3. Tabbing into or out of a cross-process iframe or
<webview> element could fire a focus event on both
the outer and inner frame, another race condition.

To solve this, we this change makes the following
changes:

1. Within a single accessibility tree, the focused node
is now part of the tree data, rather than a state flag on
a node. That guarantees only one node within a tree
can be focused at once. This affects both the web
accessibility tree and the views accessibility tree.

2. Whenever we get an accessibility focus event, we
start at the root and figure out what node has focus,
globally. Then we only fire the accessibility focus event
on that new node if focus has changed.

As an example, suppose we open a Chrome app that
embeds a <webview> that loads the google.com homepage,
which focuses the input element. Initially the app window
itself does not have focus, so we don't fire any focus events.

Later the app window gets activated. We now follow the focus
chain to the <webview> element and to the focused input
element, and fire a focus event on the input element
directly.

BUG=581910

Review URL: https://codereview.chromium.org/1589623002

Cr-Original-Commit-Position: refs/heads/master@{#376339}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 965ba000de4c91156005af7e23235574f8c2cc93