Introduce EventSourceParser
This change the parsing functionality from EventSource class. This change
improves the performance for long messages. In my environment, it reduces
the time taken for receiving a 16m long message from over 30s to under 2s.

This CL does not change the behavior, except one point. The spec says
the last event ID string should be updated when there is a valid id
field in an event even when the event is not dispatched because
the data buffer is empty. The previous implementation was wrong on this
point and this CL fixes it.

BUG=570148

Review URL: https://codereview.chromium.org/1642563002

Cr-Original-Commit-Position: refs/heads/master@{#376496}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f8259072812989d0042d6c287c03a9a351e2dfff