Remove uses of Skia's SkUtils.h.
Skia is in the process of making this header private. Most files
which currently include it are not actually using anything from it.
The one use (to check for a lead surrogate) is replaced with other
libraries more suited to the task.

Review URL: https://codereview.chromium.org/1717363002

Cr-Original-Commit-Position: refs/heads/master@{#376759}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: d2eaca13a6e6c1200b8d010c7757ccc5a4d39e43