Promote assertion after layout to RELEASE_ASSERT
We're still seeing cases of the tree stuck needing layout post-layout
in the wild. Let's fail faster to prove this is a failing of layout to
exit with the tree clean and not some other part of the system marking
it later.

BUG=545039

Review URL: https://codereview.chromium.org/1725953002

Cr-Original-Commit-Position: refs/heads/master@{#377092}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 06864c31e62c405a374c9820f76cfd81818da479