[Line Layout API] Remove LineLayoutRubyRun uses of LayoutObject* operator()
The LayoutObject* operator() is a dirty escape hatch that hides gaps in
the Line Layout API. We need to remove it before the API is done.

This CL is trivial save for making LineLayoutRubyRun a friend of
LineLayoutItem to get access to LineLayoutItem::layoutObject().

BUG=499321

Review URL: https://codereview.chromium.org/1725483002

Cr-Original-Commit-Position: refs/heads/master@{#377096}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 4938c4e383e1badc563a3f1bfff576d3239628ed