Rename OffScreenCanvas to OffscreenCanvasTemp (eventually OffscreenCanvas)
When we start it, this was named OffScreenCanvas. The whatwg proposal here:
https://wiki.whatwg.org/wiki/OffscreenCanvas actually names it OffscreenCanvas.

We cannot directly change OffScreenCanvas to OffscreenCanvas because windows
and Mac are case-insensitive. So in this CL we change it to a temp name and
in a follow up CL this will be changed to OffscreenCanvas.

Review URL: https://codereview.chromium.org/1727813002

Cr-Original-Commit-Position: refs/heads/master@{#377340}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 31d944d5d63830330f376e946b08e08347e4c22b