Fix main thread overscroll glow when scrolling iframes.
A residual scroll on a FrameView should only be handled as overscroll if the
FrameView is the main frame. This is a simple enough fix.

I also discovered we had a test for this case but it didn't catch the bug.
Further investigation showed the tests for the overscroll feature were running
in a WebView without initializing its size, leading to all sorts of accidental
behavior. I've fixed the tests now. Additionally, I've made the tests run twice,
once using touchpad scrolls and once for touchscreen to exercise both paths.

BUG=589206

Review URL: https://codereview.chromium.org/1724353002

Cr-Original-Commit-Position: refs/heads/master@{#377412}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: bc345e0531133ca5677841871bd0b116af0e53c2