Call propagateStyleToAnonymousChildren in LayoutInline::styleDidChange
propagateStyleToAnonymousChildren() is called from styleDidChange() in
LayoutBlock, LayoutTable, and a few other places, but not in
LayoutInline. Since anonymous inline-table is created for inline-level
parent (crrev.com/366564), LayoutInline not propagating to its
anonymous children will leave the style of the anonymous inline-table
not being updated, and styleDidChange() for the anonymous inline-table
not being called.

This patch adds a call to propagateStyleToAnonymousChildren() in
LayoutInline in the same way as LayoutBlock does.

One example case is when a vertical flow document is changed to
horizontal flow, the anonymous inline-table is left in vertical flow
and thus creates two orthogonal writing-mode roots without proper
notifications.

BUG=588711

Review URL: https://codereview.chromium.org/1720193004

Cr-Original-Commit-Position: refs/heads/master@{#377506}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 26d90fd22b2210d760e985c61388580eb7982044