Fix expectations on mac ports.
Apparently we append a bunch of expectations files in
base._port_specific_expectations_files(), even though they're
really not port-specific in most ways these days. So, when I
added the override for the Mac, the mac ports stopped looking
at those files, and we started getting a lot of failures as a result.

This CL fixes that, and also fixes the mock_drt implementation
to work correctly so that we can test ports w/o running them again.

I expect that as a result of this we can remove most of the 10.9
and 10.10 expectations in the Mac10_11_Expectations file.

This CL removes just a couple as an initial test.

TBR=ojan@chromium.org
BUG=589709

Review URL: https://codereview.chromium.org/1736063002 .

Cr-Original-Commit-Position: refs/heads/master@{#377666}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 68509c4733fecc5e15ae3747c72201109dc0a060