Add internals.isCSSPropertyUseCounted functionality
This patch adds the functionality of
internals.isCSSPropertyUseCounted to test whether a particular CSSProperty
is in use in a file. This function is being added to help improve testing
for improvements made to UseCounter methods for counting CSSProperty.

There is a currently a internals.isUseCounted that is used for
testing whether a current feature is UseCounted.

BUG=589343

Review URL: https://codereview.chromium.org/1711703002

Cr-Original-Commit-Position: refs/heads/master@{#377846}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 04de728dedc272262c320c5d24f5df8df2a0db7e